rewrite: complete rewrite of the operator's wiring
we're now using a debounced channel to communicate
the changed watched resources from the kubernetes watchers.

the watcher events are converted into operations that are then
applied to a state object.

when the debounce timeout kicks in the old state and the new state
are diffed to find changes.

the operator implementation should now have enough information about
nodes and services to function correctly.